added configuration changes to helmchart, addressed comments in the pr

+++ model-engine/model_engine_server/domain/use_cases/llm_model_endpoint_use_cases.py
@@ -3443,8 +3443,7 @@ def infer_addition_engine_args_from_model_name(
     return VLLMEndpointAdditionalArgs(
         gpu_memory_utilization=gpu_memory_utilization,
         attention_backend=attention_backend,
-        trust_remote_code=trust_remote_code,
-        dtype="float16",
+        trust_remote_code=trust_remote_code
     )
